requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
shap>=0.42.0
fairlearn>=0.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.28.0
imbalanced-learn>=0.10.0
lime>=0.2.0
scipy>=1.9.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Azure SDK dependencies
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-ai-ml>=1.12.0
azure-core>=1.29.0
azure-keyvault-secrets>=4.7.0
